updated db.js to export db; moved models to models folder; moved movie model to Movie.model.js
