+++ server/models/Movie.model.js
@@ -1,0 +1,36 @@
+import { DataTypes, Model } from "sequelize";
+import util from 'util';
+import { db } from "../config/db.js";
+
+export default class Movie extends Model {
+  [util.inspect.custom]() {
+    return this.toJSON();
+  }
+}
+
+Movie.init(
+  {
+    movieId: {
+      type: DataTypes.INTEGER,
+      autoIncrement: true,
+      primaryKey: true,
+    },
+    title: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    overview: {
+      type: DataTypes.TEXT,
+    },
+    releaseDate: {
+      type: DataTypes.DATE,
+    },
+    posterPath: {
+      type: DataTypes.STRING,
+    },
+  },
+  {
+    modelName: 'movie',
+    sequelize: db,
+  },
+);
