+++ server/models/index.js
@@ -1,0 +1,4 @@
+import Movie from "./Movie.model.js";
+
+
+export { Movie };
